requirements.txt
python-telegram-bot==13.14
urllib3==1.26.15
pymorphy3==2.0.2